RecommendationSvc: added recommendation

+++ CatalogSvc/Models/Product.cs
@@ -14,7 +14,7 @@ namespace CatalogSvc.Models
         public string Slug { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public float Price { get; set; }
+        public decimal Price { get; set; }
         public string Currency { get; set; }
         public string CategoryId { get; set; }
         public string CategoryName { get; set; }
